+++ components/mui-label/index.js
@@ -1,10 +1,13 @@
 import React from 'react';
 
+import { RequiredIcon } from '../required-icon';
+
 import './mui-label.scss';
 
-const MuiLabel = ({ children }) => (
+const MuiLabel = ({ children, required = false }) => (
   <div className="lf-react-material-ui-label">
     {children}
+    {required && <RequiredIcon />}
   </div>
 );